+++ conanfile.py
@@ -57,8 +57,11 @@ class OusterSDKConan(ConanFile):
 
     def requirements(self):
         # not required directly here but because libtins and libcurl pulling
-        # slightly different versions of zlib we need to set it here explicitly
+        # slightly different versions of zlib and openssl we need to set it
+        # here explicitly
         self.requires("zlib/1.2.13")
+        self.requires("openssl/1.1.1s")
+
         self.requires("eigen/3.4.0")
         self.requires("jsoncpp/1.9.5")
         self.requires("spdlog/1.10.0")